Download files in profile page
